refactor(gym): split tests_reporting.rs again (#1266)
tests_reporting.rs 752 -> 380 LOC; new tests_reporting_more.rs.
