+++ src/gym/mod.rs
@@ -19,6 +19,8 @@ mod tests_reporting;
 #[cfg(test)]
 mod tests_reporting_extra;
 #[cfg(test)]
+mod tests_reporting_more;
+#[cfg(test)]
 mod tests_scenarios;
 #[cfg(test)]
 mod tests_types;
